refactor: NicknameGenerator 위치 수정
- member의 도메인 규칙을 고려하여 NicknameGenerator의 패키지 위치를 수정
